Fix bug + light refactoring + attempt at one to one relationship in jpa
Fixed bug where multiple versions of a dependency were breaking the site.

Light refactoring for a misplaced API endpoint.

Started an attempt to get OneToOne working for census data.  Trying to solve the issue where the data is eagerly loaded when loading a CensusTract objcet. Didn't end up too well but I'm leaving the work here for now in case it's helpful in the future.