Authentication now setup
A lot of this code has to do with creating a JWT and placing in a cookie. I decided after this that the session cookie created by Spring Boot was sufficient for my needs, so the next commit you see will probably be omitting a lot of the commented out code in this.